Initial commit of reduction of memory usage

+++ CryptoGramBot/Configuration/GeneralConfig.cs
@@ -3,7 +3,7 @@
     public class GeneralConfig
     {
         public string DatabaseLocation { get; set; }
-        public int TimeOffset { get; set; }
+        public double TimeOffset { get; set; }
 
         public string TradingCurrency { get; set; }
     }
